task 12

+++ more_functions_nested_loops/101-print_number.c
@@ -16,11 +16,11 @@ void print_number(int n)
 	if (n < 0)
 	{
 		_putchar('-');
-		num = (unsigned int)(-n);
+		num = -n;
 	}
 	else
 	{
-		num = (unsigned int)(n);
+		num = n;
 	}
 
 	if (num / 10)
